Working Speech
Working speech. There are some issues with the displaying of the
results, although it is getting them just like the example given to us
in the plugin.

I'm going to work on displaying them at the top, or maybe as a menu so
that the user can pick between them.